+++ mino/springboot/Chapter05.md
@@ -1,0 +1,34 @@
+# Chapter05_프로젝트 세팅하기 - API 응답 통일, 에러 핸들러
+
+# 이슈 내용
+
+아래 6개의 과제를 수행하시면 됩니다.
+많은 과제이지만 이전 과제를 리뷰하면서 여러분의 열정이라면 충분하다고 생각합니다!
+
+## 1. 학습 후기
+
+## 2. 핵심 키워드 정리
+
+### 빌더패턴이란?
+
+### record vs static class
+
+### 제네릭이란?
+
+### @RestControllerAdvice이란?
+
+### Optional이란?
+
+
+## 3.  미션
+-  프로젝트 세팅을 마친 상태에서 응답 통일 객체, 에러 핸들링할 객체를 생성하기
+   (워크북에서 언급 안한 방식으로 생성 시, 그 이유도 함께 기재!)
+
+-  2주차 미션으로 진행한 API 명세서를 기반으로 Controller, DTO 제작하기
+   (Service & Repository는 다음 주차에서 하겠습니다!!)
+
+## 4.  워크북의 내용을 제외한 추가 정보 학습의 경우 md파일에 추가 작성합니다.
+
+## 5. 미션 제출
+- 노션에 있는 노션 페이지 게시 후 제출입니다.
+
