+++ mino/springboot/main/java/org/example/swaggerpr/mission/dto/MissionReqDto.java
@@ -1,4 +1,11 @@
 package org.example.swaggerpr.mission.dto;
 
+import lombok.Getter;
+
 public class MissionReqDto {
+
+    @Getter
+    public static class CompleteMissionDto {
+        private String status;
+    }
 }
