+++ mino/springboot/main/java/org/example/swaggerpr/review/dto/ReviewReqDto.java
@@ -1,4 +1,12 @@
 package org.example.swaggerpr.review.dto;
 
+import lombok.Getter;
+
 public class ReviewReqDto {
+
+    @Getter
+    public static class CreateReviewDto {
+        private Integer score;
+        private String content;
+    }
 }
