--- mino/springboot/main/main/java/org/example/swaggerpr/SwaggerprApplication.java
@@ -1,13 +1,0 @@
-package org.example.swaggerpr;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SwaggerprApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(SwaggerprApplication.class, args);
-    }
-
-}
