--- mino/springboot/main/main/java/org/example/swaggerpr/member/converter/MemberConverter.java
@@ -1,4 +1,0 @@
-package org.example.swaggerpr.member.converter;
-
-public class MemberConverter {
-}
