--- mino/springboot/main/main/java/org/example/swaggerpr/member/entity/FoodCategory.java
@@ -1,4 +1,0 @@
-package org.example.swaggerpr.member.entity;
-
-public class FoodCategory {
-}
