--- mino/springboot/main/main/java/org/example/swaggerpr/member/entity/Member.java
@@ -1,4 +1,0 @@
-package org.example.swaggerpr.member.entity;
-
-public class Member {
-}
