--- mino/springboot/main/main/java/org/example/swaggerpr/member/entity/mapping/MemberFoodCategory.java
@@ -1,4 +1,0 @@
-package org.example.swaggerpr.member.entity.mapping;
-
-public class MemberFoodCategory {
-}
