--- mino/springboot/main/main/java/org/example/swaggerpr/member/enums/Gender.java
@@ -1,4 +1,0 @@
-package org.example.swaggerpr.member.enums;
-
-public enum Gender {
-}
