--- mino/springboot/main/main/java/org/example/swaggerpr/member/enums/UserState.java
@@ -1,4 +1,0 @@
-package org.example.swaggerpr.member.enums;
-
-public enum UserState {
-}
