--- mino/springboot/main/main/java/org/example/swaggerpr/member/exception/MemberException.java
@@ -1,7 +1,0 @@
-package org.example.swaggerpr.member.exception;
-
-public class MemberException extends RuntimeException {
-    public MemberException(String message) {
-        super(message);
-    }
-}
