--- mino/springboot/main/main/java/org/example/swaggerpr/member/repository/MemberRepository.java
@@ -1,4 +1,0 @@
-package org.example.swaggerpr.member.repository;
-
-public interface MemberRepository {
-}
