--- mino/springboot/main/main/java/org/example/swaggerpr/member/service/MemberService.java
@@ -1,4 +1,0 @@
-package org.example.swaggerpr.member.service;
-
-public class MemberService {
-}
