--- mino/springboot/main/main/java/org/example/swaggerpr/mission/converter/MissionConverter.java
@@ -1,4 +1,0 @@
-package org.example.swaggerpr.mission.converter;
-
-public class MissionConverter {
-}
