--- mino/springboot/main/main/java/org/example/swaggerpr/mission/entity/Store.java
@@ -1,4 +1,0 @@
-package org.example.swaggerpr.mission.entity;
-
-public class Store {
-}
