--- mino/springboot/main/main/java/org/example/swaggerpr/mission/entity/mapping/MemberMission.java
@@ -1,4 +1,0 @@
-package org.example.swaggerpr.mission.entity.mapping;
-
-public class MemberMission {
-}
