--- mino/springboot/main/main/java/org/example/swaggerpr/mission/enums/Status.java
@@ -1,4 +1,0 @@
-package org.example.swaggerpr.mission.enums;
-
-public enum Status {
-}
