--- mino/springboot/main/main/java/org/example/swaggerpr/mission/exception/MissionException.java
@@ -1,7 +1,0 @@
-package org.example.swaggerpr.mission.exception;
-
-public class MissionException extends RuntimeException {
-    public MissionException(String message) {
-        super(message);
-    }
-}
