--- mino/springboot/main/main/java/org/example/swaggerpr/mission/repository/MissionRepository.java
@@ -1,4 +1,0 @@
-package org.example.swaggerpr.mission.repository;
-
-public interface MissionRepository {
-}
