--- mino/springboot/main/main/java/org/example/swaggerpr/mission/service/MissionService.java
@@ -1,4 +1,0 @@
-package org.example.swaggerpr.mission.service;
-
-public class MissionService {
-}
