--- mino/springboot/main/main/java/org/example/swaggerpr/review/converter/ReviewConverter.java
@@ -1,4 +1,0 @@
-package org.example.swaggerpr.review.converter;
-
-public class ReviewConverter {
-}
