--- mino/springboot/main/main/java/org/example/swaggerpr/review/dto/ReviewReqDto.java
@@ -1,12 +1,0 @@
-package org.example.swaggerpr.review.dto;
-
-import lombok.Getter;
-
-public class ReviewReqDto {
-
-    @Getter
-    public static class CreateReviewDto {
-        private Integer score;
-        private String content;
-    }
-}
