--- mino/springboot/main/main/java/org/example/swaggerpr/review/entity/Review.java
@@ -1,4 +1,0 @@
-package org.example.swaggerpr.review.entity;
-
-public class Review {
-}
