--- mino/springboot/main/main/java/org/example/swaggerpr/review/exception/ReviewException.java
@@ -1,7 +1,0 @@
-package org.example.swaggerpr.review.exception;
-
-public class ReviewException extends RuntimeException {
-    public ReviewException(String message) {
-        super(message);
-    }
-}
