--- mino/springboot/main/main/java/org/example/swaggerpr/review/repository/ReviewRepository.java
@@ -1,4 +1,0 @@
-package org.example.swaggerpr.review.repository;
-
-public interface ReviewRepository {
-}
