--- mino/springboot/main/main/java/org/example/swaggerpr/review/service/ReviewService.java
@@ -1,4 +1,0 @@
-package org.example.swaggerpr.review.service;
-
-public class ReviewService {
-}
