--- mino/springboot/main/main/resources/application.yml
@@ -1,19 +1,0 @@
-spring:
-  application:
-    name: "swaggerer" # "umc10th"
-
-  datasource:
-    driver-class-name: com.mysql.cj.jdbc.Driver # MySQL JDBC ???? ??? ??
-    url: ${DB_URL} # jdbc:mysql://localhost:3306/{???????}
-    username: ${DB_USER} # MySQL ?? ??
-    password: ${DB_PW} # MySQL ????
-
-  jpa:
-    database: mysql # ??? ?????? ?? ?? (MySQL)
-    database-platform: org.hibernate.dialect.MySQLDialect # Hibernate?? ??? MySQL ??(dialect) ??
-    show-sql: true # ??? SQL ??? ??? ???? ?? ??
-    hibernate:
-      ddl-auto: update # ?????? ?? ? ?????? ???? ??? ??
-    properties:
-      hibernate:
-        format_sql: true # ???? SQL ??? ?? ?? ???
